Refactor: Request Response 나누기 (#14)
